requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asn1crypto==1.5.1
blinker==1.8.2
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
keyring==24.3.1
locust==2.29.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
msgpack==1.0.8
orjson==3.10.5
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
psutil==6.0.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
requirements-parser==0.9.0
rich==13.7.1
SecretStorage==3.3.3
setuptools==70.0.0
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
snowflake-cli-labs==2.5.0
snowflake-connector-python==3.10.1
sortedcontainers==2.4.0
starlette==0.37.2
tomlkit==0.12.5
typer==0.12.3
types-setuptools==70.1.0.20240627
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
zope.event==5.0
zope.interface==6.4.post2
